+++ app/controllers/api/account.php
@@ -236,9 +236,11 @@ App::post('/v1/account/sessions')
             ->setStatusCode(Response::STATUS_CODE_CREATED)
         ;
 
+        $countries = $locale->getText('countries');
+
         $session
             ->setAttribute('current', true)
-            ->setAttribute('countryName', (isset($countries[$session->getAttribute('countryCode')])) ? $countries[$session->getAttribute('countryCode')] : $locale->getText('locale.country.unknown'))
+            ->setAttribute('countryName', (isset($countries[strtoupper($session->getAttribute('countryCode'))])) ? $countries[strtoupper($session->getAttribute('countryCode'))] : $locale->getText('locale.country.unknown'))
         ;
         
         $response->dynamic($session, Response::MODEL_SESSION);
@@ -679,8 +681,8 @@ App::get('/v1/account/sessions')
                 continue;
             }
 
-            $token->setAttribute('countryName', (isset($countries[$token->getAttribute('contryCode')]))
-                ? $countries[$token->getAttribute('contryCode')]
+            $token->setAttribute('countryName', (isset($countries[strtoupper($token->getAttribute('countryCode'))]))
+                ? $countries[strtoupper($token->getAttribute('countryCode'))]
                 : $locale->getText('locale.country.unknown'));
             $token->setAttribute('current', ($current == $token->getId()) ? true : false);