+++ CHANGES.md
@@ -6,7 +6,7 @@
 
 ## Features
 
-- Better error logs on appwrite cretificates worker
+- Better error logs on appwrite certificates worker
 - Added option for Redis authentication
 - Force adding a security email on setup
 - SMTP is now disabled by default, no dummy SMTP is included in setup
@@ -28,6 +28,7 @@
 
 - Updated missing storage env vars
 - Fixed a bug, that Response format header was not added in the access-control-allow-header list.
+- Fixed a bug where countryName is unknown on sessions (#933)
 
 ## Security